+++ CHANGELOG.md
@@ -5,6 +5,7 @@ This is the changelog for [Authress SDK](readme.md).
 * [Breaking] UserId is now required in all `userPermissions` apis. This improves **Security By Default** requiring explicit check on who the user is.
 * [Breaking] Removal of property `accessToAllSubResources`.
 * [Breaking] `ServiceClientTokenProvider` is now a first-class Javascript Class, it cannot be used as a function.
+* [Breaking] `setToken` has been removed from the interface. To set a user token, pass in a function into the AuthressClient constructor.
 
 ## 2.3 ##
 * Require minimum Node version to be 16.