+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 This is the changelog for [Authress SDK](readme.md).
 
 ## 3.0 ##
+* Require minimum Node version to be 18.
 * [Breaking] UserId is now required in all `userPermissions` apis. This improves **Security By Default** requiring explicit check on who the user is.
 * [Breaking] Removal of property `accessToAllSubResources`.
 * [Breaking] `ServiceClientTokenProvider` is now a first-class Javascript Class, it cannot be used as a function.